% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaizeGO.B73.Uniprot}
\alias{MaizeGO.B73.Uniprot}
\title{Data describing manually curated GO annotations for Maize}
\format{A data frame with 985 rows and 7 variables:
\describe{
  \item{geneID}{geneID}
  \item{goTerm}{goTerm}
  \item{publication}{publication}
  \item{evCode}{evCode}
  \item{curator}{curator}
  \item{source}{source}
  \item{type}{type}
}}
\usage{
data(MaizeGO)
}
\description{
This is data pulled from UniProt using: https://www.ebi.ac.uk/QuickGO/annotations?taxonId=4577&taxonUsage=descendants
Publication records may be DOI's instead of PMID's.  GeneID's are UniProt accessions.
}
\keyword{datasets}
