% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaizeGO.B73.v3}
\alias{MaizeGO.B73.v3}
\title{Data describing manually curated GO annotations for Maize}
\format{A data frame with 77860 rows and 7 variables:
\describe{
  \item{geneID}{geneID}
  \item{goTerm}{goTerm}
  \item{publication}{publication}
  \item{evCode}{evCode}
  \item{curator}{curator}
  \item{source}{source}
  \item{type}{type}
}}
\usage{
data(MaizeGO)
}
\description{
This is data merged from several sources including student curators,
MaizeCyc, the MaizeGDB tables, and what we call the "Gold Set".  Only
contains sources which assign GO annotations to maize B73 v3 identifiers (GRMZM's),
including the old BAC-based ids (AC's). Some GO type's are NA, possibly due to outdated terms.
}
\keyword{datasets}
