% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaizeGO.B73.v4}
\alias{MaizeGO.B73.v4}
\title{Data describing manually curated GO annotations for Maize}
\format{A data frame with 3 rows and 7 variables:
\describe{
  \item{geneID}{geneID}
  \item{goTerm}{goTerm}
  \item{publication}{publication}
  \item{evCode}{evCode}
  \item{curator}{curator}
  \item{source}{source}
  \item{type}{type}
}}
\usage{
data(MaizeGO)
}
\description{
This only contains sources that assigned GO annotations to maize B73 v4 identifiers (Zm #'s),
which currently only includes 3 assignments made by Jesse and Mary based on student work.
}
\keyword{datasets}
