% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaizeGO}
\alias{MaizeGO}
\title{Data describing manually curated GO annotations for Maize}
\format{A data frame with 3209 rows and 6 variables:
\describe{
  \item{type}{GO Annotation Data}
  \item{organism}{Zea mays}
}}
\usage{
data(MaizeGO)
}
\description{
This is data merged from several sources including student curators,
MaizeCyc, the MaizeGDB tables, UniProt, and what we call the "Gold Set"
}
\keyword{datasets}
